+++ scripts/macros/2024/spells/falseLife.js
@@ -1,0 +1,5 @@
+export let falseLife = {
+    name: 'False Life',
+    version: '1.3.156',
+    rules: 'modern'
+};